Moved config.py and serializer.py to new package.
