+++ src/QAT/qat/purr/compiler/frontends.py
@@ -3,13 +3,12 @@
 import abc
 import os
 import tempfile
-from enum import Enum
 
 import regex
 from qat.purr.backends.calibrations.remote import find_calibration
 from qat.purr.backends.realtime_chip_simulator import get_default_RTCS_hardware
 from qat.purr.compiler.builders import InstructionBuilder
-from qat.purr.compiler.config import CompilerConfig, Languages, get_optimizer_config
+from qat_config.config import CompilerConfig, Languages, get_optimizer_config
 from qat.purr.compiler.metrics import CompilationMetrics
 from qat.purr.compiler.optimisers import DefaultOptimizers
 from qat.purr.compiler.runtime import execute_instructions, get_builder, get_model
