+++ src/QAT/qat/purr/integrations/qir.py
@@ -3,7 +3,7 @@
 from typing import List, Union
 
 from qat.purr.compiler.builders import InstructionBuilder
-from qat.purr.compiler.config import InlineResultsProcessing
+from qat_config.config import InlineResultsProcessing
 from qat.purr.compiler.execution import InstructionExecutionEngine
 from qat.purr.compiler.hardware_models import QuantumHardwareModel
 from qat.purr.compiler.instructions import Variable
