+++ src/QAT/qat/purr/integrations/qiskit.py
@@ -5,7 +5,7 @@ import uuid
 from typing import List
 
 from qat.purr.backends.realtime_chip_simulator import get_default_RTCS_hardware
-from qat.purr.compiler.config import CompilerConfig
+from qat_config.config import CompilerConfig
 from qat.purr.utils.logger import get_default_logger
 from qat.qat import execute_qasm
 from qiskit import QuantumCircuit, assemble, transpile
