+++ src/QAT/qat/purr/integrations/tket.py
@@ -34,7 +34,7 @@ from pytket.passes import (
 )
 from pytket.qasm import circuit_to_qasm_str
 from pytket.qasm.qasm import NOPARAM_COMMANDS, PARAM_COMMANDS, QASMUnsupportedError
-from qat.purr.compiler.config import TketOptimizations
+from qat_config.config import TketOptimizations
 from qat.purr.compiler.execution import QuantumHardwareModel
 from qat.purr.integrations.qasm import BitRegister, Qasm2Parser, QasmContext, QubitRegister
 from qat.purr.utils.logger import get_default_logger
