+++ src/tests/test_instructions.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 from qat.purr.backends.echo import EchoEngine, get_default_echo_hardware
 from qat.purr.compiler.builders import InstructionBuilder
-from qat.purr.compiler.config import InlineResultsProcessing
+from qat_config.config import InlineResultsProcessing
 from qat.purr.compiler.devices import (
     PhysicalBaseband,
     PhysicalChannel,
