+++ src/tests/test_logger.py
@@ -11,7 +11,7 @@ import unittest
 from tempfile import gettempdir
 
 import qat.purr.utils.logger as logger
-from qat.purr.utils.serializer import json_load
+from qat_config.serializer import json_load
 
 info_msg_pattern = \
     r"^\[INFO\] \d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2},\d{3} \(\w+.\w+:\d+\) - (.+)\n"
